requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
airportsdata==20250224
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
banks==2.1.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blake3==1.0.4
bleach==6.2.0
blinker==1.9.0
boto3==1.37.16
botocore==1.37.16
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.6
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
compressed-tensors==0.9.2
contourpy==1.3.1
cryptography==44.0.2
cupy-cuda12x==13.4.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.9
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docling==2.28.4
docling-core==2.25.0
docling-ibm-models==3.4.1
docling-parse==4.0.0
durationpy==0.9
easyocr==1.7.2
einops==0.8.1
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.1.1
griffe==1.7.2
grpcio==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib-metadata==6.11.0
importlib_resources==6.5.2
interegular==0.3.3
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
langchain-huggingface==0.1.2
langchain-ollama==0.3.0
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.22
lark==1.2.2
latex2mathml==3.77.0
lazy_loader==0.4
lightning==2.5.1
lightning-cloud==0.5.70
lightning-utilities==0.14.2
lightning_sdk==0.2.4
litdata==0.2.32
litserve==0.2.6
llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index==0.12.28
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.28
llama-index-embeddings-huggingface==0.5.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.10
llama-index-llms-groq==0.3.1
llama-index-llms-langchain==0.6.1
llama-index-llms-ollama==0.5.4
llama-index-llms-openai==0.3.29
llama-index-llms-openai-like==0.3.4
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
llguidance==0.7.11
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
mistune==3.1.3
mmh3==5.1.0
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.35.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
onnxruntime==1.21.1
openai==1.70.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pdf2image==1.17.0
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
posthog==3.25.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
PyMuPDF==1.25.5
pymupdf4llm==0.0.20
pyparsing==3.2.1
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.13.0
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.14.1
rpds-py==0.23.1
rsa==4.9
rtree==1.4.0
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.11.4
semchunk==2.2.2
Send2Trash==1.8.3
sentence-transformers==4.0.1
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
simple-term-menu==1.6.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.45.3
streamlit==1.27.2
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.3.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.2.0
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
unstructured==0.17.2
unstructured-client==0.33.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
vllm==0.8.2
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.16
XlsxWriter==3.2.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
